create config folder
