+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export * from './messages';
+export * from './seo';
+export * from './site';
