+++ src/config/seo.ts
@@ -1,0 +1,9 @@
+import { siteConfig } from './site';
+
+export const seoConfig = {
+  defaultDescription: siteConfig.shortBio,
+  title: {
+    default: `${siteConfig.name} | ${siteConfig.title}`,
+    template: `%s | ${siteConfig.name}`,
+  },
+};
